refactor: ts migration for self-exclusion-footer.js
