+++ packages/account/src/Components/self-exclusion/self-exclusion-footer.tsx
