+++ packages/account/src/Components/self-exclusion/self-exclusion-inputs.jsx
@@ -5,7 +5,7 @@ import { getBrandWebsiteName, epochToMoment, toMoment, PlatformContext, isMobile
 import { Localize, localize } from '@deriv/translations';
 import { Field, useFormikContext } from 'formik';
 import SelfExclusionContext from './self-exclusion-context';
-import SelfExclusionFooter from './self-exclusion-footer.jsx';
+import SelfExclusionFooter from './self-exclusion-footer';
 
 const SectionTitle = ({ title, has_border_line }) => {
     return (
